[ 1708224 ] Ability File Tags contains extraneous items
I have added the tags indocated as being 'Ability File Tags' and restructured the rest of the page.  It appears that the extraneous tags were intended to be located on this page.  They are not the detailed instructions, but links only.

If the decision is made that they need to go, then they can be pulled easily enough.

git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@2901 ef53c05b-3124-4d59-ae7d-fe3273f42be7