[PF] Animal Archive - add Companion Archetypes
This involves some major changes to animal companions in CR, APG and UC.
It has only been tested with a low level druid at this point, so I'm not
saying that it is bug-free, only log error free with low level druids.
But since this is a lot of work I wanted to submit it so that that the
work can be checked by jenkins and others.